Put all scripts/annotate_study.py files into one
